Fix critical EIP-155 consensus bug in chainSpecificTransactionBytes
The forge agent identified a critical consensus-breaking bug:
- chainSpecificTransactionBytes parameter was still Byte, not BigInt
- This caused silent truncation for large chain IDs
- Would break replay protection and transaction verification

Fixed:
- SignedTransaction.scala: chainSpecificTransactionBytes now uses BigInt
- ECDSASignature.scala: sign method converts r, s, v to BigInt correctly
- Updated 17 test files to remove .toByte conversions
- BlockchainsConfigSpec: Fixed test that was testing the truncation bug

EIP-155 implementation now fully consensus-compliant:
- Transaction hash: keccak256(rlp([..., chainId, 0, 0]))
- V value formula: v = chainId * 2 + 35 + {0,1}
- Chain ID extraction: chainId = (v - 35) / 2
- Supports arbitrary chain IDs without truncation

Verified against ETC (chainId 61, v=157/158) and other networks.

+++ src/main/scala/com/chipprbots/ethereum/domain/SignedTransaction.scala
@@ -316,7 +316,7 @@ object SignedTransaction {
     * @return
     *   the transaction payload for Legacy transaction
     */
-  private def chainSpecificTransactionBytes(tx: Transaction, chainId: Byte): Array[Byte] = {
+  private def chainSpecificTransactionBytes(tx: Transaction, chainId: BigInt): Array[Byte] = {
     val receivingAddressAsArray: Array[Byte] = tx.receivingAddress.map(_.toArray).getOrElse(Array.emptyByteArray)
     crypto.kec256(
       rlpEncode(
