+++ src/test/scala/com/chipprbots/ethereum/blockchain/data/BootstrapCheckpointLoaderSpec.scala
@@ -138,7 +138,7 @@ class BootstrapCheckpointLoaderSpec extends AnyFlatSpec with Matchers with MockF
       customGenesisJsonOpt = None,
       daoForkConfig = None,
       accountStartNonce = com.chipprbots.ethereum.domain.UInt256.Zero,
-      chainId = 61.toByte,
+      chainId = 61,
       networkId = 1,
       monetaryPolicyConfig = MonetaryPolicyConfig(
         firstEraBlockReward = BigInt(5000000000000000000L),
