+++ src/test/scala/com/chipprbots/ethereum/consensus/pow/validators/RestrictedEthashBlockHeaderValidatorSpec.scala
@@ -106,7 +106,7 @@ class RestrictedEthashBlockHeaderValidatorSpec
         daoForkConfig = None,
         // unused
         maxCodeSize = None,
-        chainId = 0x3d.toByte,
+        chainId = 0x3d,
         networkId = 1,
         monetaryPolicyConfig = null,
         customGenesisFileOpt = None,
